Orders 7764 render create return page with context (#2673)
* feat(returns): ORDERS-7764 render create return page with stencil context

* feat(returns): ORDERS-7764 render create return page with stencil context

* feat(returns): ORDERS-7764 render create return page with stencil context

* feat(returns): ORDERS-7764 render create return page with stencil context

* feat(returns): ORDERS-7764 fix code review comments

* feat(returns): ORDERS-7764 fix code review comments

* feat(returns): ORDERS-7764 fix design feedback

* feat(returns): ORDERS-7764 fix design feedback
