+++ assets/scss/components/_components.scss
@@ -105,11 +105,14 @@
 // Write a Review
 @import "stencil/writeReview/component";
 
-// Add Return
+// Add Return (legacy)
 @import "stencil/addReturn/component";
 
 // Return Details
 @import "stencil/returnDetails/component";
+
+// Create Return
+@import "stencil/createReturn/component";
 
 // Product Carousel
 @import "stencil/productCarousel/component";
