+++ assets/scss/components/stencil/addReturn/_addReturn.scss
@@ -139,249 +139,3 @@
 }
 
 
-// =============================================================================
-// NEW RETURN PAGE (.newReturn)
-// =============================================================================
-
-.newReturn {
-    color: color("link");
-    max-width: remCalc(1100px);
-    margin: 0 auto;
-    padding: 0 spacing("single") spacing("double");
-}
-
-.newReturn-header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    margin-bottom: spacing("eighth");
-}
-
-.newReturn-titleGroup {
-    display: flex;
-    align-items: center;
-    gap: spacing("half");
-}
-
-.newReturn-title {
-    font-size: remCalc(28px);
-    font-weight: fontWeight("bold");
-    margin: 0;
-}
-
-.newReturn-statusBadge {
-    display: inline-block;
-    padding: spacing("eighth") spacing("quarter");
-    border: remCalc(1.5px) solid color("orderStatus", "completed");
-    border-radius: remCalc(4px);
-    font-size: remCalc(11px);
-    font-weight: fontWeight("bold");
-    letter-spacing: remCalc(1px);
-    color: color("orderStatus", "completed");
-    text-transform: uppercase;
-}
-
-.newReturn-headerActions {
-    display: flex;
-    align-items: center;
-    gap: spacing("half");
-}
-
-.newReturn-btnBack {
-    padding: spacing("quarter") spacing("single") * 0.83;
-    border: remCalc(1.5px) solid color("link");
-    border-radius: remCalc(6px);
-    background: color("whites", "bright");
-    font-size: fontSize("smallest");
-    font-weight: fontWeight("medium");
-    cursor: pointer;
-    text-decoration: none;
-    color: color("link");
-}
-
-.newReturn-btnPolicy {
-    padding: spacing("quarter") spacing("single") * 0.83;
-    border: 0;
-    border-radius: remCalc(6px);
-    background: color("greys", "darkest");
-    color: color("whites", "bright");
-    font-size: fontSize("smallest");
-    font-weight: fontWeight("medium");
-    cursor: pointer;
-    text-decoration: none;
-}
-
-.newReturn-orderDate {
-    font-size: remCalc(13px);
-    color: color("greys", "base");
-    margin: spacing("eighth") 0 spacing("single") * 0.83;
-}
-
-.newReturn-divider {
-    border: 0;
-    border-top: container("border");
-    margin: spacing("single") 0;
-}
-
-
-.newReturn-orderLineItem {
-    display: flex;
-    align-items: flex-start; // controls column is tallest when stacked (mobile/tablet)
-    gap: remCalc(16px);
-    padding: remCalc(16px) 0;
-
-    @include breakpoint("large") {
-        align-items: center; // desktop: everything inline so vertical-center is fine
-    }
-}
-
-.newReturn-orderLineItemThumbnailWrapper {
-    flex-shrink: 0;
-    line-height: 0; // collapse inline whitespace around the img
-}
-
-.newReturn-orderLineItemThumbnail {
-    width: remCalc(72px);
-    height: remCalc(72px);
-    object-fit: cover;
-    border-radius: remCalc(6px);
-    flex-shrink: 0;
-    background: color("greys", "lighter");
-}
-
-.newReturn-orderLineItemInfo {
-    flex: 1;
-    min-width: 0;
-}
-
-.newReturn-orderLineItemName {
-    font-size: remCalc(15px);
-    font-weight: fontWeight("semibold");
-    margin: 0 0 remCalc(2px);
-}
-
-.newReturn-orderLineItemVariant {
-    font-size: remCalc(13px);
-    color: color("greys", "base");
-    margin: 0 0 remCalc(2px);
-}
-
-.newReturn-orderLineItemPrice {
-    font-size: remCalc(13px);
-    color: color("greys", "dark");
-    margin: 0;
-}
-
-// Per-item controls cluster: stepper + resolution-select + reason-select.
-// Below large breakpoint they stack vertically (mobile + tablet); on large they
-// sit inline as in the Figma desktop frame.
-.newReturn-orderLineItemControls {
-    display: flex;
-    flex-direction: column;
-    align-items: stretch;
-    gap: spacing("quarter");
-    flex-shrink: 0;
-    width: remCalc(180px);
-
-    @include breakpoint("large") {
-        flex-direction: row;
-        align-items: center;
-        gap: spacing("half");
-        width: auto;
-    }
-}
-
-.newReturn-stepper {
-    display: inline-flex;
-    align-items: center;
-    height: remCalc(36px);
-    border: remCalc(1.5px) solid color("greys", "medium");
-    border-radius: remCalc(6px);
-    overflow: hidden;
-    flex-shrink: 0;
-    align-self: flex-start; // keep compact when stacked; selects stretch instead
-
-    @include breakpoint("large") {
-        align-self: auto;
-    }
-}
-
-.newReturn-stepperBtn {
-    background: none;
-    border: none;
-    width: remCalc(32px);
-    height: 100%;
-    font-size: remCalc(18px);
-    line-height: 1;
-    cursor: pointer;
-    color: color("greys", "dark");
-    padding: 0;
-
-    &:disabled {
-        color: color("greys", "light");
-        cursor: not-allowed;
-    }
-
-    &:focus:not(:focus-visible) {
-        outline: none !important;
-    }
-
-    &:hover:not(:disabled) {
-        background-color: color("greys", "lighter");
-    }
-}
-
-.newReturn-stepperInput {
-    width: remCalc(40px);
-    height: 100%;
-    border: none;
-    text-align: center;
-    font-size: remCalc(14px);
-    padding: 0;
-    outline: none;
-    background: transparent;
-    color: color("greys", "dark");
-}
-
-.newReturn-select {
-    width: 100%; // full width when stacked under stepper on mobile/tablet
-    height: remCalc(36px);
-    padding: 0 remCalc(28px) 0 spacing("quarter") * 1.67;
-    border: remCalc(1.5px) solid color("greys", "medium");
-    border-radius: remCalc(6px);
-    font-size: remCalc(13px);
-    color: color("greys", "dark");
-    // stylelint-disable-next-line function-url-quotes
-    background: color("whites", "bright") url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='10' height='6'%3E%3Cpath d='M0 0l5 6 5-6z' fill='%23798289'/%3E%3C/svg%3E") no-repeat right spacing("quarter") * 1.67 center;
-    background-size: remCalc(10px);
-    appearance: none;
-    cursor: pointer;
-
-    @include breakpoint("large") {
-        width: auto;
-        min-width: remCalc(180px);
-    }
-
-    &:disabled {
-        background-color: color("greys", "lighter");
-        color: color("greys", "medium");
-        cursor: not-allowed;
-        border-color: color("greys", "light");
-    }
-}
-
-// Right-aligned bottom column holding the additional-note textarea + submit button.
-// Mirrors the controls column width on tablet+; collapses to full width on mobile.
-.newReturn-footer {
-    margin-top: spacing("single");
-
-    @include breakpoint("medium") {
-        width: 50%;
-        margin-left: auto;
-    }
-
-    .form-actions {
-        text-align: right;
-    }
-}
-
