+++ assets/scss/components/stencil/createReturn/_createReturn.scss
@@ -1,0 +1,237 @@
+// =============================================================================
+// CREATE RETURN PAGE (.newReturn)
+// =============================================================================
+
+.newReturn {
+    color: color("link");
+    max-width: remCalc(1100px);
+    margin: 0 auto;
+    padding: 0 spacing("single") spacing("double");
+}
+
+.newReturn-header {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    flex-wrap: wrap;
+    gap: spacing("quarter");
+    margin-bottom: spacing("eighth");
+}
+
+.newReturn-titleGroup {
+    display: flex;
+    align-items: center;
+    gap: spacing("half");
+}
+
+.newReturn-title {
+    font-size: remCalc(28px);
+    font-weight: fontWeight("bold");
+    margin: 0;
+}
+
+.newReturn-statusBadge {
+    display: inline-block;
+    padding: spacing("eighth") spacing("quarter");
+    border: remCalc(1.5px) solid color("orderStatus", "completed");
+    border-radius: remCalc(4px);
+    font-size: remCalc(11px);
+    font-weight: fontWeight("bold");
+    letter-spacing: remCalc(1px);
+    color: color("orderStatus", "completed");
+    text-transform: uppercase;
+}
+
+.newReturn-headerActions {
+    display: flex;
+    align-items: center;
+    gap: spacing("half");
+}
+
+.newReturn-headerActions .button {
+    font-size: fontSize("smallest");
+    padding: spacing("quarter") spacing("half");
+    margin: 0;
+}
+
+.newReturn-orderDate {
+    font-size: remCalc(13px);
+    color: color("greys", "base");
+    margin: spacing("eighth") 0 spacing("single") * 0.83;
+}
+
+.newReturn-divider {
+    border: 0;
+    border-top: container("border");
+    margin: spacing("single") 0;
+}
+
+
+.newReturn-orderLineItem {
+    display: flex;
+    align-items: flex-start; // controls column is tallest when stacked (mobile/tablet)
+    gap: remCalc(16px);
+    padding: remCalc(16px) 0;
+
+    @include breakpoint("large") {
+        align-items: center; // desktop: everything inline so vertical-center is fine
+    }
+}
+
+.newReturn-orderLineItemThumbnailWrapper {
+    flex-shrink: 0;
+    line-height: 0; // collapse inline whitespace around the img
+    
+    img {
+        border: remCalc(1px) solid color("greys", "lightest");
+    }
+}
+
+.newReturn-orderLineItemThumbnail {
+    width: remCalc(72px);
+    height: remCalc(72px);
+    object-fit: cover;
+    border-radius: remCalc(6px);
+    flex-shrink: 0;
+    background: color("greys", "lighter");
+}
+
+.newReturn-orderLineItemInfo {
+    flex: 1;
+    min-width: 0;
+}
+
+.newReturn-orderLineItemName {
+    font-size: remCalc(15px);
+    font-weight: fontWeight("semibold");
+    margin: 0 0 remCalc(2px);
+}
+
+.newReturn-orderLineItemVariant {
+    font-size: remCalc(13px);
+    color: color("greys", "base");
+    margin: 0 0 remCalc(2px);
+}
+
+.newReturn-orderLineItemPrice {
+    font-size: remCalc(13px);
+    color: color("greys", "base");
+    margin: 0;
+}
+
+.newReturn-orderLineItemControls {
+    display: flex;
+    flex-direction: column;
+    align-items: stretch;
+    gap: spacing("quarter");
+    flex-shrink: 0;
+    width: remCalc(180px);
+
+    @include breakpoint("large") {
+        flex-direction: row;
+        align-items: center;
+        gap: spacing("half");
+        width: auto;
+    }
+}
+
+.newReturn-divider--headers {
+    display: none;
+
+    @include breakpoint("large") {
+        display: block;
+    }
+}
+
+.newReturn-columnHeaders {
+    display: none;
+
+    @include breakpoint("large") {
+        display: flex;
+        align-items: center;
+        gap: remCalc(16px);
+        font-size: remCalc(12px);
+        font-weight: fontWeight("semibold");
+        color: color("greys", "base");
+        letter-spacing: remCalc(0.5px);
+    }
+}
+
+// "Item" aligns with the thumbnail left edge, not the item name.
+.newReturn-columnHeaders-spacer {
+    display: none;
+}
+
+.newReturn-columnHeaders-item {
+    flex: 1;
+    min-width: 0;
+}
+
+.newReturn-columnHeaders-controls {
+    display: flex;
+    align-items: center;
+    gap: spacing("half");
+}
+
+// Shared widths — keeps header labels and control cells in sync.
+$newReturn-col-qty:     remCalc(120px);
+$newReturn-col-select:  remCalc(200px);
+
+.newReturn-columnHeaders-quantity {
+    width: $newReturn-col-qty;
+    flex-shrink: 0;
+}
+
+.newReturn-columnHeaders-request,
+.newReturn-columnHeaders-reason {
+    width: $newReturn-col-select;
+    flex-shrink: 0;
+}
+
+.newReturn-orderLineItemQtyWrapper {
+    display: flex;
+    align-items: center;
+    gap: spacing("quarter");
+
+    @include breakpoint("large") {
+        width: $newReturn-col-qty;
+        flex-shrink: 0;
+    }
+}
+
+.newReturn-orderLineItemMaxQty {
+    font-size: remCalc(13px);
+    color: color("greys", "base");
+    white-space: nowrap;
+}
+
+.newReturn-orderLineItemControls [id^="qty-"] {
+    width: 100%;
+
+    @include breakpoint("large") {
+        width: auto;
+    }
+}
+
+.newReturn-orderLineItemControls [id^="resolution-"],
+.newReturn-orderLineItemControls [id^="reason-"] {
+    width: 100%;
+
+    @include breakpoint("large") {
+        width: $newReturn-col-select;
+        flex-shrink: 0;
+    }
+}
+
+.newReturn-footer {
+    margin-top: spacing("single");
+
+    @include breakpoint("medium") {
+        width: 50%;
+        margin-left: auto;
+    }
+
+    .form-actions {
+        text-align: right;
+    }
+}
