+++ lang/en.json
@@ -340,6 +340,8 @@
                 "order_item": "Item",
                 "item_price": "Price",
                 "quantity": "Qty To Return",
+                "of_max": "of {max}",
+                "items_label": "Items in this order",
                 "why": "Why are you returning these items?",
                 "reason": "Return Reason",
                 "action": "Return Action",
@@ -446,7 +448,10 @@
             "error_no_qty": "Please select one or more items to return.",
             "none": "You haven't placed any returns with us. When you do, they will appear on this page.",
             "new_return": "New Return",
-            "from_order": "Return Items from Order #{id}",
+            "from_order": "Request Return for Order #{id}",
+            "order_date": "Order date",
+            "select_resolution": "Select preferred resolution",
+            "select_reason": "Select a return reason",
             "date_requested": "Return Requested",
             "successful_heading": "Return Request Submitted",
             "successful": "Your return was submitted successfully. We'll respond as soon as we can.",
@@ -456,10 +461,16 @@
             "action": "Return Action",
             "comments": "Your Comments",
             "submit_request": "Submit Return Request",
-            "additional_note": "Additional note",
             "additional_note_placeholder": "Please describe any additional details of this return request.",
             "submit": "Submit Return",
             "summary": "Summary",
+            "back": "Back",
+            "view_return_policy": "View return policy",
+            "additional_note": "Additional Note",
+            "request": "Preferred resolution",
+            "reason_header": "Return reason",
+            "qty": "Qty:",
+            "total_price": "Total:",
             "list": {
                 "last_update":"Last Update",
                 "return_submitted": "Return Submitted",
