--- templates/pages/account/add-return-new.html
@@ -1,114 +1,0 @@
-{{~inject 'order' orders.[0]}}
-<main class="newReturn">
-
-    {{#with orders.[0]}}
-
-    <div class="newReturn-header">
-        <div class="newReturn-titleGroup">
-            <h1 class="newReturn-title">{{lang 'account.returns.from_order' id=id}}</h1>
-            {{#if status}}<span class="newReturn-statusBadge">{{status}}</span>{{/if}}
-        </div>
-        <div class="newReturn-headerActions">
-            <a href="{{../urls.account.orders.all}}" class="newReturn-btnBack">{{lang 'account.orders.return.back_button'}}</a>
-            <a href="/return-policy" class="newReturn-btnPolicy">View return policy &rarr;</a>
-        </div>
-    </div>
-    {{#if date}}<p class="newReturn-orderDate">{{date}}</p>{{/if}}
-
-    <hr class="newReturn-divider">
-
-    <form data-new-return-form data-success-url="{{../urls.account.orders.all}}">
-
-        {{#each items}}
-            <div class="newReturn-orderLineItem" data-item-id="{{id}}">
-
-                <div class="newReturn-orderLineItemThumbnailWrapper">
-                    {{> components/common/responsive-img
-                        image=image
-                        class="newReturn-orderLineItemThumbnail"
-                        fallback_size=../../theme_settings.productthumb_size
-                        lazyload=../../theme_settings.lazyload_mode
-                        default_image=../../theme_settings.default_image_product
-                    }}
-                </div>
-
-                <div class="newReturn-orderLineItemInfo">
-                    <p class="newReturn-orderLineItemName">{{name}}</p>
-                    {{#if productAttributesList}}
-                        <p class="newReturn-orderLineItemVariant">
-                            {{#each productAttributesList}}{{name}}: {{{sanitize value}}}{{#unless @last}} &middot; {{/unless}}{{/each}}
-                        </p>
-                    {{/if}}
-                    {{!-- TODO ORDERS-7764: replace cross-ref with {{price.formatted}} directly on item once dedicated returns service includes item price --}}
-                    {{#each ../../forms.return.order_products}}
-                        {{#if id '==' ../id}}
-                            <p class="newReturn-orderLineItemPrice">{{price.formatted}} &times; {{../quantity}}</p>
-                        {{/if}}
-                    {{/each}}
-                </div>
-
-                <div class="newReturn-orderLineItemControls">
-                    <div class="newReturn-stepper">
-                        <button class="newReturn-stepperBtn" type="button"
-                                data-action="dec" disabled
-                                aria-label="Decrease quantity">&#8722;</button>
-                        {{!-- TODO ORDERS-7764: replace max={{quantity}} with returnableQuantity once dedicated returns service provides it --}}
-                        <input class="newReturn-stepperInput"
-                               id="qty-{{id}}"
-                               type="tel" value="0" min="0" max="{{quantity}}"
-                               pattern="[0-9]*" aria-label="Quantity" readonly>
-                        <button class="newReturn-stepperBtn" type="button"
-                                data-action="inc"
-                                aria-label="Increase quantity">&#43;</button>
-                    </div>
-                    <select class="newReturn-select" id="resolution-{{id}}" aria-label="Resolution dropdown">
-                        <option value="">Select a return request</option>
-                        <option value="Refund">Refund</option>
-                        <option value="Exchange">Exchange</option>
-                        <option value="Store Credit">Store Credit</option>
-                    </select>
-                    <select class="newReturn-select" id="reason-{{id}}" aria-label="Return reason">
-                        <option value="">Select a return reason</option>
-                        <option value="reason-1">Damaged or defective</option>
-                        <option value="reason-2">Wrong item received</option>
-                        <option value="reason-3">Changed my mind</option>
-                        <option value="reason-4">Item not as described</option>
-                        <option value="reason-5">Arrived too late</option>
-                    </select>
-                </div>
-            </div>
-        {{/each}}
-
-        <hr class="newReturn-divider">
-
-        <div id="return-new-error" class="alertBox alertBox--error" style="display:none" role="alert">
-            <p class="alertBox-message" data-return-error-message></p>
-        </div>
-
-        <div class="newReturn-footer">
-            <div class="form-field">
-                <label for="newReturn-additionalNote" class="form-label">
-                    {{lang 'account.returns.additional_note'}}
-                </label>
-                <textarea class="form-input"
-                          id="newReturn-additionalNote"
-                          name="additional_note"
-                          rows="4"
-                          data-new-return-note
-                          placeholder="{{lang 'account.returns.additional_note_placeholder'}}"></textarea>
-            </div>
-
-            <div class="form-actions">
-                <input class="button button--primary"
-                       id="return-new-submitBtn"
-                       type="submit"
-                       value="{{lang 'account.returns.submit'}}"
-                       disabled>
-            </div>
-        </div>
-
-    </form>
-
-    {{/with}}
-
-</main>
